+++ tests/test_core.py
@@ -8,7 +8,6 @@ import glob
 import hashlib
 import json
 import os
-import random
 import re
 import shutil
 import sys
@@ -24,7 +23,7 @@ from tools.shared import Building, STDOUT, PIPE, run_js, run_process, try_delete
 from tools.shared import NODE_JS, V8_ENGINE, JS_ENGINES, SPIDERMONKEY_ENGINE, PYTHON, EMCC, EMAR, WINDOWS, MACOS, AUTODEBUGGER
 from tools import jsrun, shared
 from runner import RunnerCore, path_from_root, core_test_modes, EMTEST_SKIP_SLOW
-from runner import skip_if, no_wasm_backend, no_fastcomp, needs_dlfcn, no_windows, env_modify, with_env_modify, is_slow_test, create_test_file
+from runner import skip_if, no_wasm_backend, no_fastcomp, needs_dlfcn, no_windows, env_modify, with_env_modify, is_slow_test, create_test_file, parameterized
 
 # decorators for limiting which modes a test can run in
 
@@ -108,6 +107,13 @@ def no_optimize(note=''):
       func(self)
     return decorated
   return decorator
+
+
+def needs_make(note=''):
+  assert not callable(note)
+  if WINDOWS:
+    return unittest.skip('Tool not available on Windows bots (%s)' % note)
+  return lambda f: f
 
 
 class TestCoreBase(RunnerCore):
@@ -5138,6 +5144,7 @@ PORT: 3979
     expected = open(path_from_root('tests', 'netinet', 'in.out')).read()
     self.do_run(src, expected)
 
+  @needs_dlfcn
   def test_main_module_static_align(self):
     if self.get_setting('ALLOW_MEMORY_GROWTH'):
       self.skipTest('no shared modules with memory growth')
@@ -5156,11 +5163,11 @@ PORT: 3979
         return 0;
       }
     '''
-    num = 5
+    num = 3
 
     def test():
       print('(iteration)')
-      time.sleep(random.random() / (10 * num)) # add some timing nondeterminism here, not that we need it, but whatever
+      time.sleep(1.0)
       self.do_run(src, 'hello world\n77.\n')
       ret = open('src.cpp.o.js', 'rb').read()
       if self.get_setting('WASM') and not self.get_setting('WASM2JS'):
@@ -5168,14 +5175,13 @@ PORT: 3979
       return ret
 
     builds = [test() for i in range(num)]
-    print(list(map(len, builds)))
+    print([len(b) for b in builds])
     uniques = set(builds)
     if len(uniques) != 1:
-      i = 0
-      for unique in uniques:
+      for i, unique in enumerate(uniques):
         open('unique_' + str(i) + '.js', 'wb').write(unique)
-        i += 1
-      assert 0, 'builds must be deterministic, see unique_X.js'
+      # builds must be deterministic, see unique_N.js
+      self.assertEqual(len(uniques), 1)
 
   def test_stdvec(self):
     self.do_run_in_out_file_test('tests', 'core', 'test_stdvec')
@@ -5511,6 +5517,7 @@ return malloc(size);
       self.emcc_args += ['-s', 'EMTERPRETIFY_WHITELIST=["_frexpl"]'] # test double call assertions
       test()
 
+  @needs_dlfcn
   def test_relocatable_void_function(self):
     self.set_setting('RELOCATABLE', 1)
     self.do_run_in_out_file_test('tests', 'core', 'test_relocatable_void_function')
@@ -5570,6 +5577,7 @@ return malloc(size);
 
     self.do_run(open(path_from_root('third_party', 'gcc_demangler.c')).read(), '*d_demangle(char const*, int, unsigned int*)*', args=['_ZL10d_demanglePKciPj'])
 
+  @needs_make('make')
   def test_lua(self):
     if self.emcc_args:
       self.emcc_args = ['-g1'] + self.emcc_args
@@ -5590,7 +5598,7 @@ return malloc(size);
                   includes=[path_from_root('tests', 'lua')],
                   output_nicerizer=lambda string, err: (string + err).replace('\n\n', '\n').replace('\n\n', '\n'))
 
-  @no_windows('./configure scripts dont to run on windows.')
+  @needs_make('configure script')
   @is_slow_test
   def test_freetype(self):
     assert 'asm2g' in core_test_modes
@@ -5658,6 +5666,7 @@ return malloc(size);
                 includes=[path_from_root('tests', 'sqlite')],
                 force_c=True)
 
+  @needs_make('mingw32-make')
   @is_slow_test
   def test_zlib(self):
     self.maybe_closure()
@@ -5682,6 +5691,7 @@ return malloc(size);
                 includes=[path_from_root('tests', 'zlib'), 'building', 'zlib'],
                 force_c=True)
 
+  @needs_make('make')
   @is_slow_test
   def test_the_bullet(self): # Called thus so it runs late in the alphabetical cycle... it is long
     self.set_setting('DEAD_FUNCTIONS', ['__ZSt9terminatev'])
@@ -5707,7 +5717,7 @@ return malloc(size);
                     includes=[path_from_root('tests', 'bullet', 'src')])
       test()
 
-  @no_windows('depends on freetype, which uses a ./configure which donsnt run on windows.')
+  @needs_make('depends on freetype')
   @is_slow_test
   def test_poppler(self):
     def test():
@@ -5740,6 +5750,7 @@ return malloc(size);
       # Make sure that DFE ends up eliminating more than 200 functions (if we can view source)
       assert (num_original_funcs - self.count_funcs('src.cpp.o.js')) > 200
 
+  @needs_make('make')
   @is_slow_test
   def test_openjpeg(self):
 
@@ -6444,7 +6455,7 @@ return malloc(size);
   def test_eval_ctors(self):
     if '-O2' not in str(self.emcc_args) or '-O1' in str(self.emcc_args):
       self.skipTest('need js optimizations')
-    if self.get_setting('WASM2JS'):
+    if not self.get_setting('WASM'):
       self.skipTest('this test uses wasm binaries')
 
     orig_args = self.emcc_args
@@ -7632,6 +7643,73 @@ extern "C" {
       self.emcc_args += ['-g']
     self.do_run(open(path_from_root('tests', 'core', 'test_ubsan_minimal_errors_same_place.c')).read(),
                 expected_output='ubsan: add-overflow\n' * 5)
+
+  @parameterized({
+    'fsanitize_undefined': (['-fsanitize=undefined'],),
+    'fsanitize_integer': (['-fsanitize=integer'],),
+    'fsanitize_overflow': (['-fsanitize=signed-integer-overflow'],),
+  })
+  @no_fastcomp('ubsan not supported on fastcomp')
+  def test_ubsan_full_overflow(self, args):
+    self.emcc_args += args
+    self.do_run(open(path_from_root('tests', 'core', 'test_ubsan_full_overflow.c')).read(),
+                expected_output=[
+      "src.cpp:3:5: runtime error: signed integer overflow: 2147483647 + 1 cannot be represented in type 'int'",
+      "src.cpp:7:7: runtime error: signed integer overflow: 2147483647 + 1 cannot be represented in type 'int'",
+    ])
+
+  @parameterized({
+    'fsanitize_undefined': (['-fsanitize=undefined'],),
+    'fsanitize_return': (['-fsanitize=return'],),
+  })
+  @no_fastcomp('ubsan not supported on fastcomp')
+  def test_ubsan_full_no_return(self, args):
+    self.emcc_args += ['-Wno-return-type'] + args
+    self.do_run(open(path_from_root('tests', 'core', 'test_ubsan_full_no_return.c')).read(),
+                expected_output='src.cpp:1:5: runtime error: execution reached the end of a value-returning function without returning a value')
+
+  @parameterized({
+    'fsanitize_undefined': (['-fsanitize=undefined'],),
+    'fsanitize_integer': (['-fsanitize=integer'],),
+    'fsanitize_shift': (['-fsanitize=shift'],),
+  })
+  @no_fastcomp('ubsan not supported on fastcomp')
+  def test_ubsan_full_left_shift(self, args):
+    self.emcc_args += args
+    self.do_run(open(path_from_root('tests', 'core', 'test_ubsan_full_left_shift.c')).read(),
+                expected_output=[
+      'src.cpp:3:5: runtime error: left shift of negative value -1',
+      "src.cpp:7:5: left shift of 16 by 29 places cannot be represented in type 'int'"
+    ])
+
+  @parameterized({
+    'fsanitize_undefined': (['-fsanitize=undefined'],),
+    'fsanitize_null': (['-fsanitize=null'],),
+  })
+  @no_fastcomp('ubsan not supported on fastcomp')
+  def test_ubsan_full_null_ref(self, args):
+    self.emcc_args += ['-std=c++11'] + args
+    self.do_run(open(path_from_root('tests', 'core', 'test_ubsan_full_null_ref.cpp')).read(),
+                expected_output=[
+      "src.cpp:3:12: runtime error: reference binding to null pointer of type 'int'",
+      "src.cpp:4:13: runtime error: reference binding to null pointer of type 'int'",
+      "src.cpp:5:14: runtime error: reference binding to null pointer of type 'int'",
+    ])
+
+  @parameterized({
+    'fsanitize_undefined': (['-fsanitize=undefined'],),
+    'fsanitize_vptr': (['-fsanitize=vptr'],),
+  })
+  @no_fastcomp('ubsan not supported on fastcomp')
+  def test_ubsan_full_static_cast(self, args):
+    self.emcc_args += args
+    self.do_run(open(path_from_root('tests', 'core', 'test_ubsan_full_static_cast.cpp')).read(),
+                expected_output=[
+      "vptr.cc:18:10: runtime error: downcast of address",
+      "which does not point to an object of type 'R'",
+      "note: object is of type 'T'",
+      "vptr for 'T'",
+    ])
 
 
 # Generate tests for everything