+++ ChangeLog.md
@@ -15,6 +15,18 @@ full changeset diff at the end of each section.
 
 Current Trunk
 -------------
+
+v1.38.34: 06/03/2019
+--------------------
+ - Add support for [undefined behavior sanitizer](https://clang.llvm.org/docs/UndefinedBehaviorSanitizer.html).
+     - This allows `emcc -fsanitize=undefined` to work. (#8651)
+     - The minimal runtime (`-fsanitize-minimal-runtime`) also works. (#8617)
+
+v1.38.33: 05/23/2019
+--------------------
+ - First release to use the new chromium build infrastructure
+   https://groups.google.com/forum/#!msg/emscripten-discuss/WhDtqVyW_Ak/8DfDnfk0BgAJ
+
  - Add `emscripten_return_address` which implements the functionality of
    gcc/clang's `__builtin_return_address`. (#8617)